reorganize code: move alignment model related stuff in alignment_model.py
